Add : Detail layout, ScrollToTop
