+++ src/components/Footer/Footer.scss
@@ -3,7 +3,5 @@
 .footer {
   @include textStyle(15px, 300);
   width: 100vw;
-  padding: 20px;
-  color: white;
-  background-color: rgba(0, 0, 0);
+  padding:10px 20px;
 }
