+++ src/components/MovieSlider/MovieSlider.scss
@@ -1,0 +1,71 @@
+@import "../../styles/mixin.scss";
+
+.slider-box {
+    margin-bottom: 40px;
+    background-color: black;
+
+    @media screen and (max-width:1200px) {
+        height: 350px;
+    }
+    @media screen and (max-width:960px) {
+        height: 500px;
+    }
+    @media screen and (max-width:768px) {
+        height: 600px;
+    }
+  
+    .slide-item {
+      position: relative;
+      .slide-content {
+        position: relative;
+        .hovered {
+          filter: brightness(50%);
+        }
+        img {
+          width: 100%;
+          transition: filter 0.3 ease;
+        }
+        .overlay {
+          position: absolute;
+          top: 0;
+          left: 0;
+          width: 100%;
+          height: 100%;
+          @include flexCenter;
+          color: white;
+          text-align: center;
+  
+          &:hover {
+            cursor: pointer;
+          }
+        }
+        .movie-info {
+          h3 {
+            margin: 0;
+            font-size: 24px;
+          }
+          p {
+            margin: 10px 0 0;
+            font-size: 16px;
+          }
+        }
+      }
+    }
+  }
+  
+  .slick-slide {
+    padding: 0 10px;
+  
+    img {
+      object-fit: cover;
+      width: 100%;
+    }
+  }
+  
+  .slick-dots {
+    li button::before {
+      border-radius: 50%;
+      background-color: tomato;
+    }
+  }
+
