+++ src/components/Nav/Nav.scss
@@ -6,7 +6,6 @@
   background-color: rgba(66, 57, 57, 0.3);
   font-size: 20px;
   padding: 20px;
-  color: white;
   position: relative;
   z-index: 1;
   min-width: 700px;
