+++ src/pages/Detail/Detail.scss
@@ -1,0 +1,47 @@
+.detail-container {
+  padding: 100px 50px 0 50px;
+
+  .navigate-btn {
+    color: tomato;
+    margin-bottom: 30px;
+    cursor: pointer;
+  }
+
+  .detail-box {
+    display: flex;
+    gap: 30px;
+
+    .detail-title {
+      width: 50%;
+
+      .title {
+        font-size: 30px;
+        margin-bottom: 20px;
+      }
+
+      .detail-img {
+        width: 100%;
+        object-fit: cover;
+      }
+    }
+
+    .detail-content {
+      width: 50%;
+      overflow: hidden;
+      border-left: 1px solid tomato;
+      padding-left: 20px;
+    }
+  }
+
+  .recommend-box {
+    border-top: 1px solid tomato;
+    padding-top: 10px;
+    margin-top: 20px;
+
+    .recommend {
+      font-size: 25px;
+      color: tomato;
+      margin-bottom: 10px;
+    }
+  }
+}
