+++ src/pages/Main/Main.js
@@ -1,8 +1,5 @@
 import "./Main.scss";
-import Slider from "react-slick";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-import { useState } from "react";
+import MovieSlider from "../../components/MovieSlider/MovieSlider";
 
 const Main = () => {
   return (
@@ -35,94 +32,6 @@ const Main = () => {
       </div>
       <MovieSlider />
     </>
-  );
-};
-
-//컴포넌트 파일로 이동 예정
-const MovieSlider = () => {
-  const [hoveredIndex, setHoveredIndex] = useState(false);  const settings = {
-    dots: true,
-    infinite: true,
-    speed: 900,
-    slidesToShow: 4,
-    slidesToScroll: 1,
-    draggable: true,
-    autoplay: true,
-    pauseOnFocus: true,
-    autoplaySpeed: 4000,
-
-    responsive: [
-      {
-        breakpoint: 960,
-        settings: {
-          slidesToShow: 3,
-        },
-      },
-      {
-        breakpoint: 768,
-        settings: {
-          slidesToShow: 2,
-        },
-      },
-    ],
-  };
-
-  //fetch요청으로 받아올 데이터 예시
-  const images = [
-    {
-      id: 1,
-      src: "https://images.unsplash.com/photo-1616530940355-351fabd9524b?q=80&w=1035&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      title: "영화 제목 1",
-      description: "영화 소개글 1",
-    },
-    {
-      id: 2,
-      src: "https://images.unsplash.com/photo-1616530940355-351fabd9524b?q=80&w=1035&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      title: "영화 제목 2",
-      description: "영화 소개글 2",
-    },
-    {
-      id: 3,
-      src: "https://images.unsplash.com/photo-1616530940355-351fabd9524b?q=80&w=1035&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      title: "영화 제목 3",
-      description: "영화 소개글 2",
-    },
-    {
-      id: 4,
-      src: "https://images.unsplash.com/photo-1616530940355-351fabd9524b?q=80&w=1035&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
-      title: "영화 제목 4",
-      description: "영화 소개글 2",
-    },
-  ];
-
-  return (
-    <div className="slider-box">
-      <Slider {...settings}>
-        {images.map((image, index) => (
-          <div className="slide-item" key={image.id}>
-            <div
-              className="slide-content"
-              onMouseEnter={() => setHoveredIndex(index)}
-              onMouseLeave={() => setHoveredIndex(false)}
-            >
-              <img
-                src={image.src}
-                alt=""
-                className={hoveredIndex === index ? "hovered" : ""}
-              />
-              {hoveredIndex === index && (
-                <div className="overlay">
-                  <div className="movie-info">
-                    <h3>{image.title}</h3>
-                    <p>{image.description}</p>
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
-        ))}
-      </Slider>
-    </div>
   );
 };
