+++ src/pages/Main/Main.scss
@@ -20,7 +20,6 @@
     position: absolute;
     top: 35%;
     left: 15%;
-    color: white;
     border-left: 1px solid white;
     padding: 0 30px;
 
@@ -40,69 +39,8 @@
   width: 100vw;
   padding: 30px;
   margin-top: -67px;
-  color: white;
 
   .watch {
     color: tomato;
-  }
-}
-
-.slider-box {
-  height: 550px;
-  overflow: hidden;
-  background-color: black;
-
-  .slide-item {
-    position: relative;
-    .slide-content {
-      position: relative;
-      .hovered {
-        filter: brightness(50%);
-      }
-      img {
-        width: 100%;
-        transition: filter 0.3 ease;
-      }
-      .overlay {
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        @include flexCenter;
-        color: white;
-        text-align: center;
-
-        &:hover {
-          cursor: pointer;
-        }
-      }
-      .movie-info {
-        h3 {
-          margin: 0;
-          font-size: 24px;
-        }
-        p {
-          margin: 10px 0 0;
-          font-size: 16px;
-        }
-      }
-    }
-  }
-}
-
-.slick-slide {
-  padding: 0 10px;
-
-  img {
-    object-fit: cover;
-    width: 100%;
-  }
-}
-
-.slick-dots {
-  li button::before {
-    border-radius: 50%;
-    background-color: tomato;
   }
 }
