+++ src/styles/common.scss
@@ -8,6 +8,7 @@
 
 body{
   background-color: black;
+  color: white;
 }
 
 button:hover {
